# perf: module token factory fast path