--- a/package.json
+++ b/package.json
@@ -58,6 +58,7 @@
     ]
   },
   "dependencies": {
+    "@node-rs/xxhash": "^1.3.0",
     "@nuxtjs/opencollective": "0.3.2",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.0",