--- a/packages/core/package.json
+++ b/packages/core/package.json
@@ -30,7 +30,6 @@
     "@nuxtjs/opencollective": "0.3.2",
     "fast-safe-stringify": "2.1.1",
     "iterare": "1.2.1",
-    "object-hash": "3.0.0",
     "path-to-regexp": "3.2.0",
     "tslib": "2.4.1",
     "uuid": "9.0.0"