--- a/packages/core/package.json
+++ b/packages/core/package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@nuxtjs/opencollective": "0.3.2",
+    "@node-rs/xxhash": "^1.3.0",
     "fast-safe-stringify": "2.1.1",
     "iterare": "1.2.1",
     "path-to-regexp": "3.2.0",