## PR Checklist
Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: https://github.com/nestjs/nest/blob/master/CONTRIBUTING.md
- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)


## PR Type
What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->
- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Other... Please describe:

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying, or link to a relevant issue. -->

This PR aims to improve the speed of `ModuleTokenFactory#create`, today, here's the profiler information of the initialization (when we initialize 10k times):

```
 [Shared libraries]:
   ticks  total  nonlib   name
  36828   80.3%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    614    1.3%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
      7    0.0%          [vdso]
      1    0.0%          /usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.28

 [JavaScript]:
   ticks  total  nonlib   name
    562    1.2%    6.7%  LazyCompile: *_object /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/object-hash/index.js:192:22
    475    1.0%    5.6%  LazyCompile: *InstanceWrapper /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/injector/instance-wrapper.js:17:16
    343    0.7%    4.1%  LazyCompile: *applyDefaults /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/object-hash/index.js:61:23
    330    0.7%    3.9%  LazyCompile: *getAllFilteredMethodNames /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/metadata-scanner.js:14:31
    314    0.7%    3.7%  LazyCompile: *reflectKeyMetadata /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/scanner.js:156:23
    304    0.7%    3.6%  LazyCompile: *_string /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/object-hash/index.js:304:22
```

> [perf-startup-10000_old_module_token_factory.js.txt](https://github.com/nestjs/nest/files/10574398/perf-startup-10000_old_module_token_factory.js.txt)

The main issues are with methods that belong to `object-hash`.

Issue Number: #10844

## What is the new behavior?

Now, I made two optimizations:

- Add a fast path when don't have metadata.
- Remove `object-hash` and perform what they will do manually.

The first optimization is the simpler one, we don't need to create an object and pass it to the `object-hash`, is way simpler to just get the main information and then generate a hash from it.

The second optimization was based on: the usage of `object-hash` is redundant, we don't need a library to traverse all the properties of the object and then generate a hash, we just need a hash, so I just serialize the entire token and then create a hash from it.

The results are these:

```
ModuleTokenFactory#create x 31,722 ops/sec ±2.45% (84 runs sampled)
BetterModuleTokenFactory#create x 356,129 ops/sec ±1.94% (84 runs sampled)

ModuleTokenFactory#create with metadata x 26,051 ops/sec ±2.36% (85 runs sampled)
BetterModuleTokenFactory#create with metadata x 126,754 ops/sec ±1.82% (89 runs sampled)

Fastest is BetterModuleTokenFactory#create
```

> [Benchmark](https://gist.github.com/H4ad/2c07b7e88f4126d49a3b6e43fa44d233)

When we don't have metadata, the improvement is 11x, with metadata the improvement is almost 5x.

Now, the profiler information looks like:

```
 [Shared libraries]:
   ticks  total  nonlib   name
  25708   80.0%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    308    1.0%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
      4    0.0%          /usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.28
      1    0.0%          [vdso]

 [JavaScript]:
   ticks  total  nonlib   name
    384    1.2%    6.3%  LazyCompile: *InstanceWrapper /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/injector/instance-wrapper.js:17:16
    328    1.0%    5.3%  LazyCompile: *getAllFilteredMethodNames /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/metadata-scanner.js:14:31
    322    1.0%    5.3%  LazyCompile: *reflectKeyMetadata /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/scanner.js:156:23
    312    1.0%    5.1%  LazyCompile: *OrdinaryGetMetadata /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/reflect-metadata/Reflect.js:600:37
    239    0.7%    3.9%  LazyCompile: *next /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/iterare/lib/iterate.js:20:9
    197    0.6%    3.2%  LazyCompile: *isMethod /home/h4ad/Projects/opensource/performance-analysis/test-performance-nestjs/node_modules/@nestjs/core/metadata-scanner.js:15:26
```

> [perf-startup-10000_better_module_token_factory.js.txt](https://github.com/nestjs/nest/files/10574397/perf-startup-10000_better_module_token_factory.js.txt)


We have more space to optimize it, like changing how the hash is generated but for now I keep it simple. 

## Does this PR introduce a breaking change?
- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->


## Other information