@H4ad 2023-02-04
Hey, reading [this article](https://chromium.googlesource.com/external/github.com/Cyan4973/xxHash/+/375d401bd4a4eba07ee75d6e627546052cb5b0ec/README.md) I found a hash function that is faster than using `sha1` called xxhash and I found a node-js binding made in rust [@node-rs/xxhash](https://www.npmjs.com/package/@node-rs/xxhash), see the results:

```
ModuleTokenFactory#create x 32,258 ops/sec ±1.88% (85 runs sampled)
BetterModuleTokenFactory#create x 369,860 ops/sec ±1.99% (84 runs sampled)
BetterModuleTokenFactoryWithNewHash#create x 728,901 ops/sec ±2.14% (84 runs sampled)
ModuleTokenFactory#create with metadata x 27,770 ops/sec ±1.35% (90 runs sampled)
BetterModuleTokenFactory#create with metadata x 125,311 ops/sec ±1.31% (88 runs sampled)
BetterModuleTokenFactoryWithNewHash#create with metadata x 159,555 ops/sec ±1.33% (87 runs sampled)
Fastest is BetterModuleTokenFactoryWithNewHash#create
```

> I also update the benchmark.ts on gist, you can test for yourself.

The output of `xxhash` function is a number, like `3687026082`.

The profiler information looks like:

Baseline:

```
 [Shared libraries]:
   ticks  total  nonlib   name
  37338   80.5%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    643    1.4%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
      7    0.0%          [vdso]
      7    0.0%          /usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.28

```

> [perf-startup-10000_baseline.txt](https://github.com/nestjs/nest/files/10608971/perf-startup-10000_baseline.txt)

With new Hash:

```
 [Shared libraries]:
   ticks  total  nonlib   name
  25158   79.7%          /home/h4ad/.asdf/installs/nodejs/16.16.0/bin/node
    169    0.5%          /usr/lib/x86_64-linux-gnu/libc-2.31.so
      8    0.0%          /usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.28
      3    0.0%          [vdso]
```

> [perf-startup-10000_module_token_factory_with_xxhash.txt](https://github.com/nestjs/nest/files/10608970/perf-startup-10000_module_token_factory_with_xxhash.txt)

Compared to the `sha1`, it didn't show too much difference in profiler information, just a little bit of improvement, but I think that's a great improvement based on the `benchmark.ts`.

I push the changes, so try and see if it's okay to introduce another dependency.