@silentroach 2023-02-07
looks like it breaks builds in alpine with

```
/usr/app/node_modules/@node-rs/xxhash/index.js:226
    throw loadError
    ^

Error: Cannot find module '@node-rs/xxhash-linux-arm64-musl'
Require stack:
- /usr/app/node_modules/@node-rs/xxhash/index.js
- /usr/app/node_modules/@nestjs/core/inspector/deterministic-uuid-registry.js
- /usr/app/node_modules/@nestjs/core/inspector/uuid-factory.js
- /usr/app/node_modules/@nestjs/core/injector/instance-wrapper.js
```