@silentroach 2023-02-07
Looks like when I install @nestjs/core on macbook it also installs xxhash for darwin platform in package-lock.json and when I run `npm ci` it will install the same package for darwin instead of new for alpine :(