@Scrib3r 2023-02-07
@silentroach
I use `node:18.14.0-alpine` with latest version of npm `v9.3.1` and it's fine for me. What've set up in your `package-lock.json`?